added transfer

+++ src/main/java/com/football_manager/dto/response/TeamDtoResponse.java
@@ -13,5 +13,5 @@ public class TeamDtoResponse {
 
     private Double balance;
 
-    private Integer commissionPercentage;
+    private Double commissionPercentage;
 }
