+++ src/main/java/com/football_manager/dto/response/TeamPlayerDtoResponse.java
@@ -14,7 +14,7 @@ public class TeamPlayerDtoResponse {
 
     private Double balance;
 
-    private Integer commissionPercentage;
+    private Double commissionPercentage;
 
     private List<PlayerResponse> players;
 }
